+++ pkg/tsdb/azuremonitor/loganalytics/azure-log-analytics-datasource_test.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"net/http"
 	"net/http/httptest"
+	"regexp"
 	"strings"
 	"testing"
 	"time"
@@ -20,11 +21,9 @@ import (
 	"github.com/grafana/grafana/pkg/tsdb/azuremonitor/types"
 )
 
-func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
-	datasource := &AzureLogAnalyticsDatasource{}
+func TestBuildLogAnalyticsQuery(t *testing.T) {
 	fromStart := time.Date(2018, 3, 15, 13, 0, 0, 0, time.UTC).In(time.Local)
 	timeRange := backend.TimeRange{From: fromStart, To: fromStart.Add(34 * time.Minute)}
-	ctx := context.Background()
 	svr := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK)
@@ -89,17 +88,21 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 		},
 	}
 
+	appInsightsRegExp, err := regexp.Compile("providers/Microsoft.Insights/components")
+	if err != nil {
+		t.Error("failed to compile reg: %w", err)
+	}
+
 	tests := []struct {
-		name                     string
-		queryModel               []backend.DataQuery
-		azureLogAnalyticsQueries []*AzureLogAnalyticsQuery
-		Err                      require.ErrorAssertionFunc
+		name                   string
+		queryModel             backend.DataQuery
+		azureLogAnalyticsQuery AzureLogAnalyticsQuery
+		Err                    require.ErrorAssertionFunc
 	}{
 		{
 			name: "Query with macros should be interpolated",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
+			queryModel: backend.DataQuery{
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"resource":     "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace",
@@ -107,18 +110,16 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 							"resultFormat": "%s",
 							"dashboardTime": false
 						}
-					}`, types.TimeSeries)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				RefID:     "A",
+				TimeRange: timeRange,
+				QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
 			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: types.TimeSeries,
-					URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
-					JSON: []byte(fmt.Sprintf(`{
+			azureLogAnalyticsQuery: AzureLogAnalyticsQuery{
+				RefID:        "A",
+				ResultFormat: dataquery.ResultFormatTimeSeries,
+				URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"resource":     "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace",
@@ -126,98 +127,88 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 							"resultFormat": "%s",
 							"dashboardTime": false
 						}
-					}`, types.TimeSeries)),
-					Query:            "Perf | where ['TimeGenerated'] >= datetime('2018-03-15T13:00:00Z') and ['TimeGenerated'] <= datetime('2018-03-15T13:34:00Z') | where ['Computer'] in ('comp1','comp2') | summarize avg(CounterValue) by bin(TimeGenerated, 34000ms), Computer",
-					Resources:        []string{"/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace"},
-					TimeRange:        timeRange,
-					QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
-					AppInsightsQuery: false,
-					DashboardTime:    false,
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				Query:            "Perf | where ['TimeGenerated'] >= datetime('2018-03-15T13:00:00Z') and ['TimeGenerated'] <= datetime('2018-03-15T13:34:00Z') | where ['Computer'] in ('comp1','comp2') | summarize avg(CounterValue) by bin(TimeGenerated, 34000ms), Computer",
+				Resources:        []string{"/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace"},
+				TimeRange:        timeRange,
+				QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
+				AppInsightsQuery: false,
+				DashboardTime:    false,
 			},
 			Err: require.NoError,
 		},
 		{
 			name: "Legacy queries with a workspace GUID should use workspace-centric url",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
+			queryModel: backend.DataQuery{
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"workspace":    "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
 							"query":        "Perf",
 							"resultFormat": "%s"
 						}
-					}`, types.TimeSeries)),
-					RefID:     "A",
-					QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				RefID:     "A",
+				QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
 			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: types.TimeSeries,
-					URL:          "v1/workspaces/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/query",
-					JSON: []byte(fmt.Sprintf(`{
+			azureLogAnalyticsQuery: AzureLogAnalyticsQuery{
+				RefID:        "A",
+				ResultFormat: dataquery.ResultFormatTimeSeries,
+				URL:          "v1/workspaces/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/query",
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"workspace":    "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
 							"query":        "Perf",
 							"resultFormat": "%s"
 						}
-					}`, types.TimeSeries)),
-					Query:            "Perf",
-					Resources:        []string{},
-					QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
-					AppInsightsQuery: false,
-					DashboardTime:    false,
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				Query:            "Perf",
+				Resources:        []string{},
+				QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
+				AppInsightsQuery: false,
+				DashboardTime:    false,
 			},
 			Err: require.NoError,
 		},
 		{
 			name: "Legacy workspace queries with a resource URI (from a template variable) should use resource-centric url",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
+			queryModel: backend.DataQuery{
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"workspace":    "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace",
 							"query":        "Perf",
 							"resultFormat": "%s"
 						}
-					}`, types.TimeSeries)),
-					RefID:     "A",
-					QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				RefID:     "A",
+				QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
 			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: types.TimeSeries,
-					URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
-					JSON: []byte(fmt.Sprintf(`{
+			azureLogAnalyticsQuery: AzureLogAnalyticsQuery{
+				RefID:        "A",
+				ResultFormat: dataquery.ResultFormatTimeSeries,
+				URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"workspace":    "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace",
 							"query":        "Perf",
 							"resultFormat": "%s"
 						}
-					}`, types.TimeSeries)),
-					Query:            "Perf",
-					Resources:        []string{},
-					QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
-					AppInsightsQuery: false,
-					DashboardTime:    false,
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				Query:            "Perf",
+				Resources:        []string{},
+				QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
+				AppInsightsQuery: false,
+				DashboardTime:    false,
 			},
 			Err: require.NoError,
 		},
 		{
 			name: "Queries with multiple resources",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
+			queryModel: backend.DataQuery{
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"resource":     "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace",
@@ -225,17 +216,15 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 							"resultFormat": "%s",
 							"dashboardTime": false
 						}
-					}`, types.TimeSeries)),
-					RefID:     "A",
-					QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				RefID:     "A",
+				QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
 			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: types.TimeSeries,
-					URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
-					JSON: []byte(fmt.Sprintf(`{
+			azureLogAnalyticsQuery: AzureLogAnalyticsQuery{
+				RefID:        "A",
+				ResultFormat: dataquery.ResultFormatTimeSeries,
+				URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"resource":     "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace",
@@ -243,21 +232,19 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 							"resultFormat": "%s",
 							"dashboardTime": false
 						}
-					}`, types.TimeSeries)),
-					Query:            "Perf",
-					Resources:        []string{"/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace"},
-					QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
-					AppInsightsQuery: false,
-					DashboardTime:    false,
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				Query:            "Perf",
+				Resources:        []string{"/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace"},
+				QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
+				AppInsightsQuery: false,
+				DashboardTime:    false,
 			},
 			Err: require.NoError,
 		},
 		{
 			name: "Query with multiple resources",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
+			queryModel: backend.DataQuery{
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"resources":     ["/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace",  "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace2"],
@@ -265,18 +252,16 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 							"resultFormat": "%s",
 							"dashboardTime": false
 						}
-					}`, types.TimeSeries)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				RefID:     "A",
+				TimeRange: timeRange,
+				QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
 			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: types.TimeSeries,
-					URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
-					JSON: []byte(fmt.Sprintf(`{
+			azureLogAnalyticsQuery: AzureLogAnalyticsQuery{
+				RefID:        "A",
+				ResultFormat: dataquery.ResultFormatTimeSeries,
+				URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"resources":     ["/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace",  "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace2"],
@@ -284,22 +269,20 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 							"resultFormat": "%s",
 							"dashboardTime": false
 						}
-					}`, types.TimeSeries)),
-					Query:            "Perf",
-					Resources:        []string{"/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace", "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace2"},
-					TimeRange:        timeRange,
-					QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
-					AppInsightsQuery: false,
-					DashboardTime:    false,
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				Query:            "Perf",
+				Resources:        []string{"/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace", "/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace2"},
+				TimeRange:        timeRange,
+				QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
+				AppInsightsQuery: false,
+				DashboardTime:    false,
 			},
 			Err: require.NoError,
 		},
 		{
 			name: "Query that uses dashboard time",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
+			queryModel: backend.DataQuery{
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"resources":     ["/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace"],
@@ -308,18 +291,16 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 							"dashboardTime": true,
 							"timeColumn":	"TimeGenerated"
 						}
-					}`, types.TimeSeries)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				RefID:     "A",
+				TimeRange: timeRange,
+				QueryType: string(dataquery.AzureQueryTypeAzureLogAnalytics),
 			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: types.TimeSeries,
-					URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
-					JSON: []byte(fmt.Sprintf(`{
+			azureLogAnalyticsQuery: AzureLogAnalyticsQuery{
+				RefID:        "A",
+				ResultFormat: dataquery.ResultFormatTimeSeries,
+				URL:          "v1/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace/query",
+				JSON: []byte(fmt.Sprintf(`{
 						"queryType": "Azure Log Analytics",
 						"azureLogAnalytics": {
 							"resources":     ["/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace"],
@@ -328,1121 +309,14 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 							"dashboardTime": true,
 							"timeColumn":	"TimeGenerated"
 						}
-					}`, types.TimeSeries)),
-					Query:            "Perf",
-					Resources:        []string{"/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace"},
-					TimeRange:        timeRange,
-					QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
-					AppInsightsQuery: false,
-					DashboardTime:    true,
-					TimeColumn:       "TimeGenerated",
-				},
-			},
-			Err: require.NoError,
-		},
-
-		{
-			name: "trace query",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s",
-								"traceTypes":	["trace"],
-								"operationId":	"test-op-id"
-							}
-						}`, dataquery.ResultFormatTable)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTable,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s",
-								"traceTypes":	["trace"],
-								"operationId":	"test-op-id"
-							}
-						}`, dataquery.ResultFormatTable)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true trace` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true trace` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true trace` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"test-op-id\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with no result format set",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"traceTypes":	["trace"],
-								"operationId":	"test-op-id"
-							}
-						}`),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTable,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"traceTypes":	["trace"],
-								"operationId":	"test-op-id"
-							}
-						}`),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true trace` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true trace` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true trace` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"test-op-id\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with no operation ID",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTable)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTable,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTable)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == '${__data.fields.traceID}') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == '${__data.fields.traceID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == '${__data.fields.traceID}') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == '${__data.fields.traceID}')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"${__data.fields.traceID}\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with no types",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s",
-								"operationId":	"test-op-id"
-							}
-						}`, dataquery.ResultFormatTable)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTable,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s",
-								"operationId":	"test-op-id"
-							}
-						}`, dataquery.ResultFormatTable)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"test-op-id\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with eq filter",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-								"queryType": "Azure Traces",
-								"azureTraces": {
-									"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-									"resultFormat": "%s",
-									"operationId":	"test-op-id",
-									"filters":		[{"filters": ["test-app-id"], "property": "appId", "operation": "eq"}]
-								}
-							}`, dataquery.ResultFormatTable)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTable,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-								"queryType": "Azure Traces",
-								"azureTraces": {
-									"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-									"resultFormat": "%s",
-									"operationId":	"test-op-id",
-									"filters":		[{"filters": ["test-app-id"], "property": "appId", "operation": "eq"}]
-								}
-							}`, dataquery.ResultFormatTable)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId in ("test-app-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId in ("test-app-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId in ("test-app-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"test-op-id\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with ne filter",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-								"queryType": "Azure Traces",
-								"azureTraces": {
-									"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-									"resultFormat": "%s",
-									"operationId":	"test-op-id",
-									"filters":		[{"filters": ["test-app-id"], "property": "appId", "operation": "ne"}]
-								}
-							}`, dataquery.ResultFormatTable)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTable,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-								"queryType": "Azure Traces",
-								"azureTraces": {
-									"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-									"resultFormat": "%s",
-									"operationId":	"test-op-id",
-									"filters":		[{"filters": ["test-app-id"], "property": "appId", "operation": "ne"}]
-								}
-							}`, dataquery.ResultFormatTable)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId !in ("test-app-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId !in ("test-app-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId !in ("test-app-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"test-op-id\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with multiple filters",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-								"queryType": "Azure Traces",
-								"azureTraces": {
-									"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-									"resultFormat": "%s",
-									"operationId":	"test-op-id",
-									"filters":		[{"filters": ["test-app-id"], "property": "appId", "operation": "ne"},{"filters": ["test-client-id"], "property": "clientId", "operation": "eq"}]
-							}
-						}`, dataquery.ResultFormatTable)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTable,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-								"queryType": "Azure Traces",
-								"azureTraces": {
-									"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-									"resultFormat": "%s",
-									"operationId":	"test-op-id",
-									"filters":		[{"filters": ["test-app-id"], "property": "appId", "operation": "ne"},{"filters": ["test-client-id"], "property": "clientId", "operation": "eq"}]
-							}
-						}`, dataquery.ResultFormatTable)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId !in ("test-app-id")| where clientId in ("test-client-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId !in ("test-app-id")| where clientId in ("test-client-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,traces` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| where appId !in ("test-app-id")| where clientId in ("test-client-id")` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"test-op-id\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with trace result format",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTrace,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":     ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests` +
-						`| where (operation_Id != '' and operation_Id == '${__data.fields.traceID}') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == '${__data.fields.traceID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests` +
-						`| where (operation_Id != '' and operation_Id == '${__data.fields.traceID}') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == '${__data.fields.traceID}')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"${__data.fields.traceID}\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with trace result format and operation ID",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"test-op-id",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTrace,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"test-op-id",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests` +
-						`| where (operation_Id != '' and operation_Id == 'test-op-id') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'test-op-id')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName` +
-						`| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"test-op-id\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with trace result format and only trace type",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"test-op-id",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s",
-								"traceTypes":		["traces"]
-							}
-						}`, dataquery.ResultFormatTrace)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTrace,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"test-op-id",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s",
-								"traceTypes":		["traces"]
-							}
-						}`, dataquery.ResultFormatTrace)),
-					Query:                   "",
-					Resources:               []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange:               timeRange,
-					QueryType:               dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery:       "",
-					TraceParentExploreQuery: "",
-					TraceLogsExploreQuery: "union availabilityResults,\n" + "customEvents,\n" + "dependencies,\n" + "exceptions,\n" + "pageViews,\n" + "requests,\n" + "traces\n" +
-						"| where operation_Id == \"test-op-id\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with operation ID and correlated workspaces",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"op-id-multi",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTrace,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"op-id-multi",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-multi') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-multi')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-multi') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-multi')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-multi') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-multi')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union *,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').traces\n" +
-						"| where operation_Id == \"op-id-multi\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with multiple resources",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTrace,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| where (operation_Id != '' and operation_Id == '${__data.fields.traceID}') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == '${__data.fields.traceID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| where (operation_Id != '' and operation_Id == '${__data.fields.traceID}') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == '${__data.fields.traceID}')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union *,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').traces\n" +
-						"| where operation_Id == \"${__data.fields.traceID}\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with multiple resources and overlapping correlated workspaces",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"op-id-multi",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTrace,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"op-id-multi",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-multi') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-multi')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-multi') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-multi')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-multi') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-multi')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union *,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').traces\n" +
-						"| where operation_Id == \"op-id-multi\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
-			},
-			Err: require.NoError,
-		},
-		{
-			name: "trace query with multiple resources and non-overlapping correlated workspaces",
-			queryModel: []backend.DataQuery{
-				{
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"op-id-non-overlapping",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					RefID:     "A",
-					TimeRange: timeRange,
-					QueryType: string(dataquery.AzureQueryTypeAzureTraces),
-				},
-			},
-			azureLogAnalyticsQueries: []*AzureLogAnalyticsQuery{
-				{
-					RefID:        "A",
-					ResultFormat: dataquery.ResultFormatTrace,
-					URL:          "v1/apps/r1/query",
-					JSON: []byte(fmt.Sprintf(`{
-							"queryType": "Azure Traces",
-							"azureTraces": {
-								"operationId": 	"op-id-non-overlapping",
-								"resources":    ["/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"],
-								"resultFormat": "%s"
-							}
-						}`, dataquery.ResultFormatTrace)),
-					Query: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-non-overlapping') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-non-overlapping')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					Resources: []string{"/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r1", "/subscriptions/test-sub/resourceGroups/test-rg/providers/Microsoft.Insights/components/r2"},
-					TimeRange: timeRange,
-					QueryType: dataquery.AzureQueryTypeAzureTraces,
-					TraceExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-non-overlapping') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-non-overlapping')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceParentExploreQuery: `set truncationmaxrecords=10000; set truncationmaxsize=67108864; union isfuzzy=true availabilityResults,customEvents,dependencies,exceptions,pageViews,requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').availabilityResults,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').customEvents,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').dependencies,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').exceptions,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').pageViews,app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').requests` +
-						`| where (operation_Id != '' and operation_Id == 'op-id-non-overlapping') or (customDimensions.ai_legacyRootId != '' and customDimensions.ai_legacyRootId == 'op-id-non-overlapping')` +
-						`| where (operation_ParentId != '' and operation_ParentId == '${__data.fields.parentSpanID}')` +
-						`| extend duration = iff(isnull(column_ifexists("duration", real(null))), toreal(0), column_ifexists("duration", real(null)))` +
-						`| extend spanID = iff(itemType == "pageView" or isempty(column_ifexists("id", "")), tostring(new_guid()), column_ifexists("id", ""))` +
-						`| extend operationName = iff(isempty(column_ifexists("name", "")), column_ifexists("problemId", ""), column_ifexists("name", ""))` +
-						`| extend serviceName = cloud_RoleName| extend serviceTags = bag_pack_columns(cloud_RoleInstance, cloud_RoleName)` +
-						`| extend error = todynamic(iff(itemType == "exception", "true", "false"))` +
-						`| extend tags = bag_merge(bag_pack_columns(appId,appName,application_Version,assembly,client_Browser,client_City,client_CountryOrRegion,client_IP,client_Model,client_OS,client_StateOrProvince,client_Type,data,details,duration,error,handledAt,iKey,id,innermostAssembly,innermostMessage,innermostMethod,innermostType,itemCount,itemId,itemType,location,message,method,name,operation_Id,operation_Name,operation_ParentId,operation_SyntheticSource,outerAssembly,outerMessage,outerMethod,outerType,performanceBucket,problemId,resultCode,sdkVersion,session_Id,severityLevel,size,source,success,target,timestamp,type,url,user_AccountId,user_AuthenticatedId,user_Id), customDimensions, customMeasurements)` +
-						`| project-rename traceID = operation_Id, parentSpanID = operation_ParentId, startTime = timestamp` +
-						`| project startTime, itemType, serviceName, duration, traceID, spanID, parentSpanID, operationName, serviceTags, tags, itemId` +
-						`| order by startTime asc`,
-					TraceLogsExploreQuery: "union *,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').availabilityResults,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').customEvents,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').dependencies,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').exceptions,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').pageViews,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').requests,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r2').traces,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').availabilityResults,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').customEvents,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').dependencies,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').exceptions,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').pageViews,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').requests,\n" +
-						"app('/subscriptions/test-sub/resourcegroups/test-rg/providers/microsoft.insights/components/r3').traces\n" +
-						"| where operation_Id == \"op-id-non-overlapping\"",
-					AppInsightsQuery: true,
-					DashboardTime:    true,
-					TimeColumn:       "timestamp",
-				},
+					}`, dataquery.ResultFormatTimeSeries)),
+				Query:            "Perf",
+				Resources:        []string{"/subscriptions/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/resourceGroups/cloud-datasources/providers/Microsoft.OperationalInsights/workspaces/AppInsightsTestDataWorkspace"},
+				TimeRange:        timeRange,
+				QueryType:        dataquery.AzureQueryTypeAzureLogAnalytics,
+				AppInsightsQuery: false,
+				DashboardTime:    true,
+				TimeColumn:       "TimeGenerated",
 			},
 			Err: require.NoError,
 		},
@@ -1450,9 +324,9 @@ func TestBuildingAzureLogAnalyticsQueries(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			queries, err := datasource.buildQueries(ctx, tt.queryModel, dsInfo)
+			query, err := buildLogAnalyticsQuery(tt.queryModel, dsInfo, appInsightsRegExp)
 			tt.Err(t, err)
-			if diff := cmp.Diff(tt.azureLogAnalyticsQueries[0], queries[0]); diff != "" {
+			if diff := cmp.Diff(&tt.azureLogAnalyticsQuery, query); diff != "" {
 				t.Errorf("Result mismatch (-want +got): \n%s", diff)
 			}
 		})
